+++ Src/DocumentInput.cs
@@ -21,12 +21,6 @@ using MongoDB.Bson;
 
 namespace Mdbc
 {
-	interface IDocumentInput
-	{
-		PSObject InputObject { get; }
-		PSObject Id { get; }
-		SwitchParameter NewId { get; }
-	}
 	class DocumentInput
 	{
 		SessionState Session;
@@ -57,31 +51,27 @@ namespace Mdbc
 				return result.Count == 0 ? BsonNull.Value : null;
 			}
 		}
-		public static void MakeId(BsonDocument document, IDocumentInput input, SessionState session)
+		public static BsonDocument NewDocumentWithId(bool newId, PSObject id, PSObject input, SessionState session)
 		{
-			//! NewId first
-			if (input.NewId)
-			{
-				document["_id"] = new BsonObjectId(ObjectId.GenerateNewId());
-			}
-			else if (input.Id != null)
-			{
-				var sb = input.Id.BaseObject as ScriptBlock;
-				if (sb == null)
-				{
-					document["_id"] = BsonValue.Create(input.Id.BaseObject);
-				}
-				else
-				{
-					using (new SetDollar(session, input.InputObject))
-					{
-						var arr = sb.Invoke();
-						if (arr.Count != 1)
-							throw new ArgumentException("-Id script must return a single object."); //! use this type
+			if (newId && id != null) throw new PSInvalidOperationException("Parameters Id and NewId cannot be used together.");
+			
+			if (newId)
+				return new BsonDocument().Add("_id", new BsonObjectId(ObjectId.GenerateNewId()));
 
-						document["_id"] = BsonValue.Create(arr[0].BaseObject);
-					}
-				}
+			if (id == null)
+				return null;
+
+			var sb = id.BaseObject as ScriptBlock;
+			if (sb == null)
+				return new BsonDocument().Add("_id", BsonValue.Create(id.BaseObject));
+
+			using (new SetDollar(session, input))
+			{
+				var arr = sb.Invoke();
+				if (arr.Count != 1)
+					throw new ArgumentException("-Id script must return a single object."); //! use this type
+
+				return new BsonDocument().Add("_id", BsonValue.Create(arr[0].BaseObject));
 			}
 		}
 		public static ErrorRecord NewErrorRecordBsonValue(Exception value, object targetObject)